+++ routes/general.routes.js
@@ -3,10 +3,12 @@ const router = express.Router();
 
 const beneficiary_routes = require('./beneficiary.routes');
 const suppliesRoutes = require("./supplies.routes");
+const externalCollabsRoutes = require("./externalCollabs.routes");
 const supplyBatchRouter = require("./supplyBatch.routes")
 
 router.use("/beneficiary", beneficiary_routes);
 router.use("/supplies", suppliesRoutes);
+router.use("/externalCollabs", externalCollabsRoutes);
 router.use("/supplyBatch", supplyBatchRouter);